project: fix Windows builds, disabled VS2015 support

+++ src/utils/math.h
@@ -18,9 +18,9 @@ inline float lerp(float a, float b, float t) { return a + t * (b - a); }
 // Convert normalized float to int16_t
 inline int16_t floatToInt(float val) {
     if (val >= 0) {
-        return val * static_cast<int16_t>(INT16_MAX);
+        return static_cast<int16_t>(val * static_cast<int16_t>(INT16_MAX));
     } else {
-        return -val * static_cast<int16_t>(INT16_MIN);
+        return static_cast<int16_t>(-val * static_cast<int16_t>(INT16_MIN));
     }
 }
